Data.Massiv.Array.Delayed

Description

Delayed

Delayed Pull Array

data D

Delayed representation.

Constructors

D

delay :: (Index ix, Source r e) => Array r ix e -> Array D ix e

O(1) Conversion from a source array to D representation.

liftArray2' :: (HasCallStack, Index ix, Source r1 a, Source r2 b) => (a -> b -> e) -> Array r1 ix a -> Array r2 ix b -> Array D ix e

Same as liftArray2M, but throws an imprecise exception on mismatched sizes.

Since: 1.0.0

liftArray2M :: (Index ix, Source r1 a, Source r2 b, MonadThrow m) => (a -> b -> e) -> Array r1 ix a -> Array r2 ix b -> m (Array D ix e)

Similar to zipWith, except dimensions of both arrays have to be the same, otherwise it throws SizeMismatchException.

Since: 1.0.0

Delayed Push Array

data DL

Delayed load representation. Also known as Push array.

Constructors

DL

toLoadArray :: forall r ix e. (Size r, Load r ix e) => Array r ix e -> Array DL ix e

Convert any Loadable array into DL representation.

Since: 0.3.0

makeLoadArrayS

Arguments

:: forall ix e. Index ix

=> Sz ix
Size of the resulting array

-> e
Default value to use for all cells that might have been ommitted by the writing function

-> (forall m. Monad m => (ix -> e -> m Bool) -> m ())
Writing function that described which elements to write into the target array.

-> Array DL ix e

Describe how an array should be loaded into memory sequentially. For parallelizable version see makeLoadArray.

Since: 0.3.1

makeLoadArray

Arguments

:: forall ix e. Index ix

=> Comp
Computation strategy to use. Directly affects the scheduler that gets created for the loading function.

-> Sz ix
Size of the resulting array

-> e
Default value to use for all cells that might have been ommitted by the writing function

-> (forall s. Scheduler s () -> (ix -> e -> ST s Bool) -> ST s ())
Writing function that described which elements to write into the target array. It accepts a scheduler, that can be used for parallelization, as well as a safe element writing function.

-> Array DL ix e

Specify how an array should be loaded into memory. Unlike makeLoadArrayS, loading function accepts a scheduler, thus can be parallelized. If you need an unsafe version of this function see unsafeMakeLoadArray.

Since: 0.4.0

fromStrideLoad :: forall r ix e. StrideLoad r ix e => Stride ix -> Array r ix e -> Array DL ix e

Convert an array that can be loaded with stride into DL representation.

Since: 0.3.0

Delayed Stream Array

data DS

Delayed stream array that represents a sequence of values that can be loaded sequentially. Important distinction from other arrays is that its size might no be known until it is computed.

Constructors

DS

toStreamArray :: (Index ix, Source r e) => Array r ix e -> Vector DS e

Flatten an array into a stream of values.

Since: 0.4.1

toSteps :: Vector DS e -> Steps Id e

O(1) - Convert delayed stream array into Steps.

Since: 0.4.1

fromSteps :: Steps Id e -> Vector DS e

O(1) - Convert Steps into delayed stream array

Since: 0.4.1

Delayed Interleaved Array

data DI

Delayed array that will be loaded in an interleaved fashion during parallel computation.

Warning - Will be deprecated in the next major version update.

Constructors

DI

toInterleaved :: (Index ix, Source r e) => Array r ix e -> Array DI ix e

Convert a source array into an array that, when computed, will have its elemets evaluated out of order (interleaved amongst cores), hence making unbalanced computation better parallelizable.

fromInterleaved :: Array DI ix e -> Array D ix e

O(1) - Unwrap the interleved array.

Since: 0.2.1

Delayed Windowed Array

data DW

Delayed Windowed Array representation.

Constructors

DW

data Window ix e

Constructors

Window

Fields

windowStart :: !ix

Index of where window will start at.

windowSize :: !(Sz ix)

Size of the window

windowIndex :: ix -> e

Indexing function for the window

windowUnrollIx2 :: !(Maybe Int)

Setting this value during stencil application improves cache utilization by unrolling the loop for Ix2 and higher dimensions. Has no affect on arrays with one dimension.

insertWindow

Arguments

:: Index ix

=> Array D ix e
Source array that will have a window inserted into it

-> Window ix e
Window to place inside the delayed array

-> Array DW ix e

Inserts a Window into a delayed array while scaling the window down if it doesn't fit inside that array.

Since: 0.3.0

getWindow :: Array DW ix e -> Maybe (Window ix e)

Get the Window from a windowed array.

Since: 0.2.1

dropWindow :: Array DW ix e -> Array D ix e

Drop the Window from a windowed array.

Since: 0.3.0

makeWindowedArray

Arguments

:: (Index ix, Source r e)

=> Array r ix e
Source array that will have a window inserted into it

-> ix
Start index for the window

-> Sz ix
Size of the window

-> (ix -> e)
Indexing function foto use inside window

-> Array DW ix e

Construct a delayed windowed array by supply a separate element producing function for the interior of an array. This is very usful for stencil mapping, where interior function does not perform boundary checks, thus significantly speeding up computation process.

Since: 0.1.3