Data.Massiv.Array.Manifest

Description

Manifest

class Source r e => Manifest r e

Manifest arrays are backed by actual memory and values are looked up versus computed as it is with delayed arrays. Because manifest arrays are located in memory their contents can be mutated once thawed into MArray. The process of changed a mutable MArray back into an immutable Array is called freezing.

Minimal complete definition

unsafeLinearIndexM, sizeOfMArray, unsafeResizeMArray, unsafeLinearSliceMArray, unsafeThaw, unsafeFreeze, unsafeNew, unsafeLinearRead, unsafeLinearWrite, initialize

Generate

generateArray :: forall r ix e m. (MonadUnliftIO m, Manifest r e, Index ix) => Comp -> Sz ix -> (ix -> m e) -> m (Array r ix e)

Just like generateArrayS, except this generator will respect the supplied computation strategy, and for that reason it is restricted to IO.

Since: 0.2.6

generateArrayLinear :: forall r ix e m. (MonadUnliftIO m, Manifest r e, Index ix) => Comp -> Sz ix -> (Ix1 -> m e) -> m (Array r ix e)

Just like generateArray, except generating action will receive a row-major linear index.

Since: 0.3.0

generateArrayS

Arguments

:: forall r ix e m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the array

-> (ix -> m e)
Element producing action

-> m (Array r ix e)

Sequentially generate a pure array. Much like makeArray creates a pure array this function will use Manifest interface to generate a pure Array in the end, except that computation strategy is set to Seq. Element producing function no longer has to be pure but is a stateful action, becuase it is restricted to PrimMonad thus allows for sharing the state between computation of each element.

Examples

Expand

>>> import Data.Massiv.Array
>>> import Data.IORef
>>> ref <- newIORef (0 :: Int)
>>> generateArrayS (Sz1 6) (\ i -> modifyIORef' ref (+i) >> print i >> pure i) :: IO (Array U Ix1 Int)
0
1
2
3
4
5
Array U Seq (Sz1 6)
  [ 0, 1, 2, 3, 4, 5 ]
>>> readIORef ref
15

Since: 0.2.6

generateArrayLinearS

Arguments

:: forall r ix e m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Resulting size of the array

-> (Int -> m e)
Element producing generator

-> m (Array r ix e)

Same as generateArray but with action that accepts row-major linear index.

Since: 0.3.0

generateSplitSeedArray

Arguments

:: forall r ix e g it. (Iterator it, Manifest r e, Index ix)

=> it
Iterator

-> g
Initial seed

-> (forall s. g -> ST s (g, g))
An ST action that can split a seed into two independent seeds. It will be called the same number of times as the number of jobs that will get scheduled during parallelization. Eg. only once for the sequential case.

-> Comp
Computation strategy.

-> Sz ix
Resulting size of the array.

-> (forall s. Ix1 -> ix -> g -> ST s (e, g))
An ST action that produces a value and the next seed. It takes both versions of the index, in linear and in multi-dimensional forms, as well as the current seeding value. Returns the element for the array cell together with the new seed that will be used for the next element generation

-> (g, [g], Array r ix e)
Returned values are:

The final split of the supplied seed.

Results of scheduled jobs in the same order that they where scheduled

Final array that was fully filled using the supplied action and iterator.

Similar to makeSplitSeedArray, except it will produce a Manifest array and will return back the last unused seed together with all final seeds produced by each scheduled job. This function can be thought of as an unfolding done in parallel while iterating in a customizable manner.

Since: 1.0.2

Stateful worker threads

generateArrayWS :: forall r ix e s m. (Manifest r e, Index ix, MonadUnliftIO m, PrimMonad m) => WorkerStates s -> Sz ix -> (ix -> s -> m e) -> m (Array r ix e)

Use per worker thread state while generating elements of the array. Very useful for things that are not thread safe.

Since: 0.3.4

generateArrayLinearWS :: forall r ix e s m. (Manifest r e, Index ix, MonadUnliftIO m, PrimMonad m) => WorkerStates s -> Sz ix -> (Int -> s -> m e) -> m (Array r ix e)

Same as generateArrayWS, but use linear indexing instead.

Since: 0.3.4

Unfold

unfoldrPrimM_

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> m (e, a))
Unfolding action

-> a
Initial accumulator

-> m (Array r ix e)

Sequentially unfold an array from the left.

Examples

Expand

Create an array with Fibonacci numbers while performing an IO action at each iteration.

>>> import Data.Massiv.Array
>>> unfoldrPrimM_ (Sz1 10) (\(f0, f1) -> (f0, (f1, f0 + f1)) <$ print f1) (0, 1) :: IO (Array P Ix1 Int)
1
1
2
3
5
8
13
21
34
55
Array P Seq (Sz1 10)
  [ 0, 1, 1, 2, 3, 5, 8, 13, 21, 34 ]

Since: 0.3.0

iunfoldrPrimM_

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> ix -> m (e, a))
Unfolding action

-> a
Initial accumulator

-> m (Array r ix e)

Same as unfoldrPrimM_ but do the unfolding with index aware function.

Since: 0.3.0

unfoldrPrimM

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> m (e, a))
Unfolding action

-> a
Initial accumulator

-> m (a, Array r ix e)

Just like iunfoldrPrimM, but do the unfolding with index aware function.

Since: 0.3.0

iunfoldrPrimM

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> ix -> m (e, a))
Unfolding action

-> a
Initial accumulator

-> m (a, Array r ix e)

Just like iunfoldrPrimM_, but also returns the final value of the accumulator.

Since: 0.3.0

unfoldlPrimM_

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> m (a, e))
Unfolding action

-> a
Initial accumulator

-> m (Array r ix e)

Sequentially unfold an array from the left.

Examples

Expand

Create an array with Fibonacci numbers starting at the end while performing and IO action on the accumulator for each element of the array.

>>> import Data.Massiv.Array
>>> unfoldlPrimM_ (Sz1 10) (\a@(f0, f1) -> let fn = f0 + f1 in print a >> return ((f1, fn), f0)) (0, 1) :: IO (Array P Ix1 Int)
(0,1)
(1,1)
(1,2)
(2,3)
(3,5)
(5,8)
(8,13)
(13,21)
(21,34)
(34,55)
Array P Seq (Sz1 10)
  [ 34, 21, 13, 8, 5, 3, 2, 1, 1, 0 ]

Since: 0.3.0

iunfoldlPrimM_

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> ix -> m (a, e))
Unfolding action

-> a
Initial accumulator

-> m (Array r ix e)

Same as unfoldlPrimM_ but do the unfolding with index aware function.

Since: 0.3.0

unfoldlPrimM

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> m (a, e))
Unfolding action

-> a
Initial accumulator

-> m (a, Array r ix e)

Just like iunfoldlPrimM, but do the unfolding with index aware function.

Since: 0.3.0

iunfoldlPrimM

Arguments

:: forall r ix e a m. (Manifest r e, Index ix, PrimMonad m)

=> Sz ix
Size of the desired array

-> (a -> ix -> m (a, e))
Unfolding action

-> a
Initial accumulator

-> m (a, Array r ix e)

Just like iunfoldlPrimM_, but also returns the final value of the accumulator.

Since: 0.3.0

Mapping

forPrimM :: (Manifest r e, Index ix, PrimMonad m) => MArray (PrimState m) r ix e -> (e -> m e) -> m ()

Sequentially loop over a mutable array while modifying each element with an action.

Since: 0.4.0

forPrimM_ :: (Manifest r e, Index ix, PrimMonad m) => MArray (PrimState m) r ix e -> (e -> m ()) -> m ()

Sequentially loop over a mutable array while reading each element and applying an action to it. There is no mutation to the array, unless the action itself modifies it.

Since: 0.4.0

iforPrimM :: (Manifest r e, Index ix, PrimMonad m) => MArray (PrimState m) r ix e -> (ix -> e -> m e) -> m ()

Sequentially loop over a mutable array while modifying each element with an index aware action.

Since: 0.4.0

iforPrimM_ :: (Manifest r e, Index ix, PrimMonad m) => MArray (PrimState m) r ix e -> (ix -> e -> m ()) -> m ()

Sequentially loop over a mutable array while reading each element and applying an index aware action to it. There is no mutation to the array, unless the action itself modifies it.

Since: 0.4.0

iforLinearPrimM :: (Manifest r e, Index ix, PrimMonad m) => MArray (PrimState m) r ix e -> (Int -> e -> m e) -> m ()

Sequentially loop over a mutable array while modifying each element with an index aware action.

Since: 0.4.0

iforLinearPrimM_ :: (Manifest r e, Index ix, PrimMonad m) => MArray (PrimState m) r ix e -> (Int -> e -> m ()) -> m ()

Sequentially loop over a mutable array while reading each element and applying a linear index aware action to it. There is no mutation to the array, unless the action itself modifies it.

Since: 0.4.0

for2PrimM_ :: forall r1 r2 e1 e2 ix m. (PrimMonad m, Index ix, Manifest r1 e1, Manifest r2 e2) => MArray (PrimState m) r1 ix e1 -> MArray (PrimState m) r2 ix e2 -> (e1 -> e2 -> m ()) -> m ()

Sequentially loop over the intersection of two mutable arrays while reading elements from both and applying an action to it. There is no mutation to the actual arrays, unless the action itself modifies either one of them.

Since: 1.0.0

ifor2PrimM_ :: forall r1 r2 e1 e2 ix m. (PrimMonad m, Index ix, Manifest r1 e1, Manifest r2 e2) => MArray (PrimState m) r1 ix e1 -> MArray (PrimState m) r2 ix e2 -> (ix -> e1 -> e2 -> m ()) -> m ()

Same as for2PrimM_, but with index aware action.

Since: 1.0.0

Boxed

data B

Array representation for Boxed elements. Its elements are strict to Weak Head Normal Form (WHNF) only.

Constructors

B

data BL

Array representation for Boxed elements. This data structure is lazy with respect to its elements.

Example

Expand

Memoized version of a factorial that relies on laziness. Note that computing memoized factorial of a million would likely overflow memory.

>>> import Data.Massiv.Array as A
>>> :{
mkMemoFactorial :: Int -> (Int -> Integer)
mkMemoFactorial n =
  let arr = makeVectorR BL Seq (Sz1 n) fact
      fact i | i == 0 = 1
             | otherwise = (arr ! (i - 1)) * toInteger i
  in (arr !)
:}

>>> let fact = mkMemoFactorial 1000001
>>> fact 50
30414093201713378043612608166064768844377641568960512000000000000
>>> length $ show $ fact 5000
16326

Constructors

BL

data BN

Constructors

BN

type N = BN

Deprecated: In favor of more consistently named BN

Type and pattern N have been added for backwards compatibility and will be replaced in the future in favor of BN.

Deprecated - since 1.0.0

pattern N :: N

Deprecated: In favor of more consistently named BN

data Uninitialized

An error that gets thrown when an unitialized element of a boxed array gets accessed. Can only happen when array was constructed with unsafeNew.

Constructors

Uninitialized

Access

findIndex :: (Index ix, Manifest r e) => (e -> Bool) -> Array r ix e -> Maybe ix

O(n) - Perform a row-major search starting at 0 for an element. Returns the index of the first occurance of an element or Nothing if a predicate could not be satisifed after it was applyied to all elements of the array.

Since: 0.5.5

Conversion

Important part of all conversions in this section is that the actual boxed Array, which holds the pointers to values isn't copied around, it is always kept as the same array. Conversion to Massiv boxed array will undergo evaluation during which computation strategies will be respected.

toLazyArray :: Array B ix e -> Array BL ix e

O(1) - Cast a strict boxed array into a lazy boxed array.

Since: 0.6.0

evalLazyArray :: Index ix => Array BL ix e -> Array B ix e

O(n) - Evaluate all elements of a boxed lazy array to weak head normal form

Since: 0.6.0

forceLazyArray :: (NFData e, Index ix) => Array BL ix e -> Array N ix e

O(n) - Evaluate all elements of a boxed lazy array to normal form

Since: 0.6.0

unwrapNormalForm :: Array N ix e -> Array B ix e

O(1) - Converts array from N to B representation.

Since: 0.5.0

evalNormalForm :: (Index ix, NFData e) => Array B ix e -> Array N ix e

O(n) - Compute all elements of a boxed array to NF (normal form)

Since: 0.5.0

Primitive Boxed Array

unwrapLazyArray :: Array BL ix e -> Array e

O(1) - Unwrap boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.6.0

wrapLazyArray :: Array e -> Vector BL e

O(1) - Wrap a boxed array.

Since: 0.6.0

unwrapArray :: Array B ix e -> Array e

O(1) - Unwrap boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

evalArray

Arguments

:: Comp
Computation strategy

-> Array e
Lazy boxed array from primitive package.

-> Vector B e

O(n) - Wrap a boxed array and evaluate all elements to a WHNF.

Since: 0.2.1

unwrapMutableArray :: MArray s B ix e -> MutableArray s e

O(1) - Unwrap mutable boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

unwrapMutableLazyArray :: MArray s BL ix e -> MutableArray s e

O(1) - Unwrap mutable boxed lazy array. This will discard any possible slicing that has been applied to the array.

Since: 0.6.0

evalMutableArray

Arguments

:: PrimMonad m

=> MutableArray (PrimState m) e
Mutable array that will get wrapped

-> m (MArray (PrimState m) B Ix1 e)

O(n) - Wrap mutable boxed array and evaluate all elements to WHNF.

Since: 0.2.1

unwrapNormalFormArray :: Array N ix e -> Array e

O(1) - Unwrap a fully evaluated boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

evalNormalFormArray

Arguments

:: NFData e

=> Comp
Computation strategy

-> Array e
Lazy boxed array

-> Array N Ix1 e

O(n) - Wrap a boxed array and evaluate all elements to a Normal Form (NF).

Since: 0.2.1

unwrapNormalFormMutableArray :: MArray s N ix e -> MutableArray s e

O(1) - Unwrap a fully evaluated mutable boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

evalNormalFormMutableArray :: (PrimMonad m, NFData e) => MutableArray (PrimState m) e -> m (MArray (PrimState m) N Ix1 e)

O(n) - Wrap mutable boxed array and evaluate all elements to NF.

Since: 0.2.1

Boxed Vector

toBoxedVector :: Index ix => Array BL ix a -> Vector a

O(1) - Converts a boxed Array into a Vector without touching any elements.

Since: 0.5.0

toBoxedMVector :: Index ix => MArray s BL ix a -> MVector s a

O(1) - Converts a boxed MArray into a MVector.

Since: 0.5.0

fromBoxedVector :: Vector a -> Vector BL a

O(1) - Cast a boxed vector without touching any elements.

Since: 0.6.0

fromBoxedMVector :: MVector s a -> MArray s BL Ix1 a

O(1) - Convert mutable boxed vector to a lazy mutable boxed array. Both keep pointing to the same memory

Since: 0.6.0

evalBoxedVector :: Comp -> Vector a -> Array B Ix1 a

O(n) - Convert a boxed vector and evaluate all elements to WHNF. Computation strategy will be respected during evaluation

Since: 0.5.0

evalBoxedMVector :: PrimMonad m => MVector (PrimState m) a -> m (MArray (PrimState m) B Ix1 a)

O(n) - Convert mutable boxed vector and evaluate all elements to WHNF sequentially. Both keep pointing to the same memory

Since: 0.5.0

Primitive

data P

Representation for Primitive elements

Constructors

P

class Prim a

Class of types supporting primitive array operations. This includes interfacing with GC-managed memory (functions suffixed with ByteArray#) and interfacing with unmanaged memory (functions suffixed with Addr#). Endianness is platform-dependent.

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

Conversion

Primitive ByteArray

toByteArray :: (Index ix, Prim e) => Array P ix e -> ByteArray

O(n) - Ensure that the size matches the internal ByteArray. If not make a copy of the slice and return it as ByteArray

Since: 0.2.1

toByteArrayM :: (Prim e, Index ix, MonadThrow m) => Array P ix e -> m ByteArray

O(1) - Unwrap Ensure that the size matches the internal ByteArray.

Since: 0.5.0

unwrapByteArray :: Array P ix e -> ByteArray

O(1) - Extract the internal ByteArray. This will ignore any possible slicing that has been applied to the array. Use toByteArray in order to preserve slicing or unwrapByteArrayOffset to get ahold of the offset

Since: 0.5.0

unwrapByteArrayOffset :: Array P ix e -> Int

O(1) - Extract potential linear offset into the underlying ByteArray, which can also be extracted with unwrapByteArray.

Since: 0.5.9

fromByteArray :: forall e. Prim e => Comp -> ByteArray -> Array P Ix1 e

O(1) - Construct a flat Array from ByteArray

Since: 0.4.0

fromByteArrayM :: (MonadThrow m, Index ix, Prim e) => Comp -> Sz ix -> ByteArray -> m (Array P ix e)

O(1) - Construct a primitive array from the ByteArray. Will return Nothing if number of elements doesn't match.

Since: 0.3.0

fromByteArrayOffsetM :: (MonadThrow m, Index ix, Prim e) => Comp -> Sz ix -> Int -> ByteArray -> m (Array P ix e)

O(1) - Construct a primitive array from the ByteArray. Will return Nothing if number of elements doesn't match.

Since: 0.5.9

toMutableByteArray :: forall ix e m. (Prim e, Index ix, PrimMonad m) => MArray (PrimState m) P ix e -> m (Bool, MutableByteArray (PrimState m))

O(n) - Try to cast a mutable array to MutableByteArray, if sizes do not match make a copy. Returns True if an array was converted without a copy, in which case it means that the source at the resulting array are still pointing to the same location in memory.

Since: 0.5.0

unwrapMutableByteArray :: MArray s P ix e -> MutableByteArray s

O(1) - Extract the internal MutableByteArray. This will discard any possible slicing that has been applied to the array.

Since: 0.5.0

unwrapMutableByteArrayOffset :: MArray s P ix e -> Int

O(1) - Extract the linear offset into underlying MutableByteArray, which can aslo be extracted with unwrapMutableByteArray.

Since: 0.5.9

fromMutableByteArray :: forall e s. Prim e => MutableByteArray s -> MArray s P Ix1 e

O(1) - Construct a flat Array from MutableByteArray

Since: 0.4.0

fromMutableByteArrayM :: (MonadThrow m, Index ix, Prim e) => Sz ix -> MutableByteArray s -> m (MArray s P ix e)

O(1) - Construct a primitive mutable array from the MutableByteArray. Will throw SizeElementsMismatchException if number of elements doesn't match.

Since: 0.3.0

fromMutableByteArrayOffsetM :: (MonadThrow m, Index ix, Prim e) => Sz ix -> Ix1 -> MutableByteArray s -> m (MArray s P ix e)

O(1) - Construct a primitive mutable array from the MutableByteArray. Will throw SizeElementsMismatchException if number of elements doesn't match.

Since: 0.5.9

Primitive Vector

toPrimitiveVector :: Index ix => Array P ix e -> Vector e

O(1) - Cast a primitive array to a primitive vector.

Since: 0.5.0

toPrimitiveMVector :: Index ix => MArray s P ix e -> MVector s e

O(1) - Cast a mutable primitive array to a mutable primitive vector.

Since: 0.5.0

fromPrimitiveVector :: Vector e -> Array P Ix1 e

O(1) - Cast a primitive vector to a primitive array.

Since: 0.5.0

fromPrimitiveMVector :: MVector s e -> MArray s P Ix1 e

O(1) - Cast a mutable primitive vector to a mutable primitive array.

Since: 0.5.0

Storable

data S

Representation for Storable elements

Constructors

S

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

mallocCompute :: forall r ix e. (Size r, Load r ix e, Storable e) => Array r ix e -> IO (Array S ix e)

Very similar to computeAs S except load the source array into memory allocated with malloc on C heap. It can potentially be useful when iteroperating with some C programs.

Since: 0.5.9

mallocCopy :: forall ix e. (Index ix, Storable e) => Array S ix e -> IO (Array S ix e)

Allocate memory on C heap with malloc and copy the source array over.

Since: 0.5.9

Conversion

Storable Vector

toStorableVector :: Index ix => Array S ix e -> Vector e

O(1) - Unwrap storable array and pull out the underlying storable vector.

Since: 0.2.1

toStorableMVector :: Index ix => MArray s S ix e -> MVector s e

O(1) - Unwrap storable mutable array and pull out the underlying storable mutable vector.

Since: 0.2.1

fromStorableVector :: Comp -> Vector e -> Vector S e

O(1) - Cast a storable vector to a storable array.

Since: 0.5.0

fromStorableMVector :: MVector s e -> MVector s S e

O(1) - Cast a mutable storable vector to a mutable storable array.

Since: 0.5.0

Direct Pointer Access

withPtr :: MonadUnliftIO m => MArray RealWorld S ix e -> (Ptr e -> m b) -> m b

A pointer to the beginning of the mutable array.

Since: 0.1.3

Unboxed

data U

Representation for Unboxed elements

Constructors

U

class (Vector Vector a, MVector MVector a) => Unbox a

Conversion

Unboxed Vector

toUnboxedVector :: Array U ix e -> Vector e

O(1) - Unwrap unboxed array and pull out the underlying unboxed vector.

Since: 0.2.1

toUnboxedMVector :: MArray s U ix e -> MVector s e

O(1) - Unwrap unboxed mutable array and pull out the underlying unboxed mutable vector.

Since: 0.2.1

fromUnboxedVector :: Unbox e => Comp -> Vector e -> Vector U e

O(1) - Wrap an unboxed vector and produce an unboxed flat array.

Since: 0.6.0

fromUnboxedMVector :: Unbox e => MVector s e -> MVector s U e

O(1) - Wrap an unboxed mutable vector and produce a mutable unboxed flat array.

Since: 0.5.0

ByteString Conversion

fromByteString

Arguments

:: Load r Ix1 Word8

=> Comp
Computation strategy

-> ByteString
Strict ByteString to use as a source.

-> Vector r Word8

O(n) - Convert a strict ByteString into a manifest array. Will return Nothing if length doesn't match the total number of elements of new array.

Since: 0.2.1

castFromByteString :: Comp -> ByteString -> Vector S Word8

O(1) - Cast a strict ByteString into a Storable array

Since: 0.3.0

toByteString

Arguments

:: Load r ix Word8

=> Array r ix Word8
Source array

-> ByteString

O(n) - Convert any source array into a strict ByteString. In case when the source array is actually storable, no memory copy will occur.

Since: 0.2.1

castToByteString :: Index ix => Array S ix Word8 -> ByteString

O(1) - Cast a Storable array into a strict ByteString

Since: 0.3.0

toBuilder :: (Index ix, Source r e) => (e -> Builder) -> Array r ix e -> Builder

O(n) - Conversion of array monoidally into a ByteString Builder.

Since: 0.2.1

castToBuilder :: Index ix => Array S ix Word8 -> Builder

O(1) - Cast a storable array of Word8 to ByteString Builder.

Since: 0.5.0