Data.Massiv.Array.Mutable.Atomic

Description

Atomic element-wise mutation

atomicReadIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> m (Maybe Int)

Atomically read an Int element from the array

Since: 0.3.0

atomicWriteIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m Bool

Atomically write an Int element int the array. Returns True if supplied index was correct and write was successfull.

Since: 0.3.0

atomicModifyIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> (Int -> Int) -> m (Maybe Int)

Atomically modify an Int element of the array. Returns the old value, unless the supplied index was out of bounds.

Since: 0.3.0

atomicAddIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m (Maybe Int)

Atomically add to an Int element in the array. Returns the old value.

Since: 0.3.0

atomicSubIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m (Maybe Int)

Atomically subtract from an Int element in the array. Returns the old value.

Since: 0.3.0

atomicAndIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m (Maybe Int)

Atomically AND an Int element in the array. Returns the old value.

Since: 0.3.0

atomicNandIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m (Maybe Int)

Atomically NAND an Int element in the array. Returns the old value.

Since: 0.3.0

atomicOrIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m (Maybe Int)

Atomically OR an Int element in the array. Returns the old value.

Since: 0.3.0

atomicXorIntArray :: (Index ix, PrimMonad m) => MArray (PrimState m) P ix Int -> ix -> Int -> m (Maybe Int)

Atomically XOR an Int element in the array. Returns the old value.

Since: 0.3.0

casIntArray

Arguments

:: (Index ix, PrimMonad m)

=> MArray (PrimState m) P ix Int
Array to mutate

-> ix
Index at which to mutate

-> Int
Expected value

-> Int
New value

-> m (Maybe Int)

Atomically CAS (Compare-and-Swap) an Int in the array. Returns the old value.

Since: 0.3.0