Data.Massiv.Core

Description

Documentation

data family Array r ix e :: Type

The array family. Representations r describe how data is arranged or computed. All arrays have a common property that each index ix always maps to the same unique element e, even if that element does not yet exist in memory and the array has to be computed in order to get the value of that element. Data is always arranged in a nested row-major fashion. Rank of an array is specified by Dimensions ix.

Since: 0.1.0

newtype List ix e

Constructors

List

Fields

unList :: [Elt ix e]

type Vector r e = Array r Ix1 e

Type synonym for a single dimension array, or simply a flat vector.

Since: 0.5.0

type MVector s r e = MArray s r Ix1 e

Type synonym for a single dimension mutable array, or simply a flat mutable vector.

Since: 0.5.0

type Matrix r e = Array r Ix2 e

Type synonym for a two-dimentsional array, or simply a matrix.

Since: 0.5.0

type MMatrix s r e = MArray s r Ix2 e

Type synonym for a two-dimentsional mutable array, or simply a mutable matrix.

Since: 0.5.0

class (Strategy r, Shape r ix) => Load r ix e where

Any array that can be computed and loaded into memory

Minimal complete definition

(makeArray | makeArrayLinear), (iterArrayLinearST_ | iterArrayLinearWithSetST_)

Methods

iterArrayLinearST_

Arguments

:: Scheduler s ()

-> Array r ix e
Array that is being loaded

-> (Int -> e -> ST s ())
Function that writes an element into target array

-> ST s ()

Iterate over an array with a ST action that is applied to each element and its index.

Since: 1.0.0

iterArrayLinearWithSetST_

Arguments

:: Scheduler s ()

-> Array r ix e
Array that is being loaded

-> (Ix1 -> e -> ST s ())
Function that writes an element into target array

-> (Ix1 -> Sz1 -> e -> ST s ())
Function that efficiently sets a region of an array to the supplied value target array

-> ST s ()

Similar to iterArrayLinearST_. Except it also accepts a function that is potentially optimized for setting many cells in a region to the same value.

Since: 1.0.0

class Load r ix e => Stream r ix e where

Methods

toStream :: Array r ix e -> Steps Id e

toStreamIx :: Array r ix e -> Steps Id (ix, e)

class (Strategy r, Size r) => Source r e

Arrays that can be used as source to practically any manipulation function.

Minimal complete definition

(unsafeIndex | unsafeLinearIndex), unsafeLinearSlice

data PrefIndex ix e

Prefered indexing function.

Constructors

PrefIndex (ix -> e)

PrefIndexLinear (Int -> e)

class Size r

Arrays that have information about their size availible in constant time.

Minimal complete definition

size, unsafeResize

class Index ix => Shape r ix where

The shape of an array. It is different from Size in that it can be applicable to non-square matrices and might not be available in constant time.

Since: 1.0.0

Minimal complete definition

Nothing

Methods

linearSizeHint :: Array r ix e -> LengthHint

O(1) - Check what do we know about the number of elements without doing any work

Since: 1.0.0

linearSize :: Array r ix e -> Sz1

O(n) - possibly iterate over the whole array before producing the answer

Since: 0.5.8

default linearSize :: Size r => Array r ix e -> Sz1

outerSize :: Array r ix e -> Sz ix

O(n) - Rectangular size of an array that is inferred from looking at the first row in each dimensions. For rectangular arrays this is the same as size

Since: 1.0.0

default outerSize :: Size r => Array r ix e -> Sz ix

maxLinearSize :: Array r ix e -> Maybe Sz1

O(1) - Get the possible maximum linear size of an immutabe array. If the lookup of size in constant time is not possible, Nothing will be returned. This value will be used as the initial size of the mutable array into which the loading will happen.

Since: 1.0.0

isNull :: Array r ix e -> Bool

O(1) - Check whether an array is empty or not.

Examples

Expand

>>> import Data.Massiv.Array
>>> isNull $ range Seq (Ix2 10 20) (11 :. 21)
False
>>> isNull $ range Seq (Ix2 10 20) (10 :. 21)
True
>>> isNull (empty :: Array D Ix5 Int)
True
>>> isNull $ sfromList []
True

Since: 1.0.0

data LengthHint

Size hint

Since: 1.0.0

Constructors

LengthExact Sz1
Exact known size

LengthMax Sz1
Upper bound on the size

LengthUnknown
Unknown size

class Load r ix e => StrideLoad r ix e where

Minimal complete definition

Nothing

Methods

iterArrayLinearWithStrideST_

Arguments

:: Scheduler s ()

-> Stride ix
Stride to use

-> Sz ix
Size of the target array affected by the stride.

-> Array r ix e
Array that is being loaded

-> (Int -> e -> ST s ())
Function that writes an element into target array

-> ST s ()

Load an array into memory with stride. Default implementation requires an instance of Source.

default iterArrayLinearWithStrideST_ :: Source r e => Scheduler s () -> Stride ix -> Sz ix -> Array r ix e -> (Int -> e -> ST s ()) -> ST s ()

class Source r e => Manifest r e

Manifest arrays are backed by actual memory and values are looked up versus computed as it is with delayed arrays. Because manifest arrays are located in memory their contents can be mutated once thawed into MArray. The process of changed a mutable MArray back into an immutable Array is called freezing.

Minimal complete definition

unsafeLinearIndexM, sizeOfMArray, unsafeResizeMArray, unsafeLinearSliceMArray, unsafeThaw, unsafeFreeze, unsafeNew, unsafeLinearRead, unsafeLinearWrite, initialize

type Mutable r e = Manifest r e

Deprecated: In favor of Manifest

Starting with massiv-1.0 Mutable and Manifest are synonymous. Since massiv-1.1 it is deprecated and will be removed in massiv-1.2

class (IsList (Array r ix e), Load r ix e) => Ragged r ix e

Minimal complete definition

generateRaggedM, flattenRagged, loadRaggedST, raggedFormat

data L

Constructors

L

type family ListItem ix e :: Type where ...

Equations

ListItem Ix1 e = e

ListItem ix e = [ListItem (Lower ix) e]

data Scheduler s a

Main type for scheduling work. See withScheduler or withScheduler_ for ways to construct and use this data type.

Since: scheduler-1.0.0

data SchedulerWS ws a

This is a wrapper around Scheduler, but it also keeps a separate state for each individual worker. See withSchedulerWS or withSchedulerWS_ for ways to construct and use this data type.

Since: scheduler-1.4.0

class Typeable r => Strategy r

Minimal complete definition

setComp, getComp

data Comp

Computation strategy to use when scheduling work.

Constructors

Seq
Sequential computation

ParOn ![Int]
Schedule workers to run on specific capabilities. Specifying an empty list ParOn [] or using Par will result in utilization of all available capabilities.

ParN !Word16
Specify the number of workers that will be handling all the jobs. Difference from ParOn is that workers can jump between cores. Using ParN 0 will result in using all available capabilities.

Bundled Patterns

pattern Par :: Comp
Parallel computation using all available cores. Same as ParOn []

Since: scheduler-1.0.0

pattern Par' :: Comp
Parallel computation using all available cores. Same as ParN 0

Since: scheduler-1.1.0

getComp :: Strategy r => Array r ix e -> Comp

Get computation strategy of this array

Since: 0.1.0

setComp :: Strategy r => Comp -> Array r ix e -> Array r ix e

Set computation strategy for this array

Example

Expand

>>> :set -XTypeApplications
>>> import Data.Massiv.Array
>>> a = singleton @DL @Ix1 @Int 0
>>> a
Array DL Seq (Sz1 1)
  [ 0 ]
>>> setComp (ParN 6) a -- use 6 capabilities
Array DL (ParN 6) (Sz1 1)
  [ 0 ]

appComp :: Strategy r => Comp -> Array r ix e -> Array r ix e

Append computation strategy using Comp's Monoid instance.

Since: 0.6.0

data WorkerStates ws

Each worker is capable of keeping it's own state, that can be share for different schedulers, but not at the same time. In other words using the same WorkerStates on withSchedulerS concurrently will result in an error. Can be initialized with initWorkerStates

Since: scheduler-1.4.0

initWorkerStates :: MonadIO m => Comp -> (WorkerId -> m ws) -> m (WorkerStates ws)

Initialize a separate state for each worker.

Since: scheduler-1.4.0

scheduleWork :: MonadPrimBase s m => Scheduler s a -> m a -> m ()

Schedule an action to be picked up and computed by a worker from a pool of jobs. Similar to scheduleWorkId, except the job doesn't get the worker id.

Since: scheduler-1.0.0

scheduleWork_ :: MonadPrimBase s m => Scheduler s () -> m () -> m ()

Same as scheduleWork, but only for a Scheduler that doesn't keep the results.

Since: scheduler-1.1.0

module Data.Massiv.Core.Index

Numeric

class (Size r, Num e) => FoldNumeric r e

Minimal complete definition

foldArray, powerSumArray, unsafeDotProduct

class FoldNumeric r e => Numeric r e

Minimal complete definition

unsafeLiftArray, unsafeLiftArray2

class (Numeric r e, Floating e) => NumericFloat r e

Exceptions

class Monad m => MonadThrow (m :: Type -> Type) where

A class for monads in which exceptions may be thrown.

Instances should obey the following law:

throwM e >> x = throwM e

In other words, throwing an exception short-circuits the rest of the monadic computation.

Methods

throwM :: (HasCallStack, Exception e) => e -> m a

Throw an exception. Note that this throws when this action is run in the monad m, not when it is applied. It is a generalization of Control.Exception's throwIO.

Should satisfy the law:

throwM e >> f = throwM e

data IndexException where

Exceptions that get thrown when there is a problem with an index, size or dimension.

Since: 0.3.0

Constructors

IndexZeroException :: Index ix => !ix -> IndexException
Index contains a zero value along one of the dimensions.

IndexDimensionException :: (NFData ix, Eq ix, Show ix, Typeable ix) => !ix -> !Dim -> IndexException
Dimension is out of reach.

IndexOutOfBoundsException :: Index ix => !(Sz ix) -> !ix -> IndexException
Index is out of bounds.

data SizeException where

Exception that indicates an issue with an array size.

Since: 0.3.0

Constructors

SizeMismatchException :: Index ix => !(Sz ix) -> !(Sz ix) -> SizeException
Two sizes are expected to be equal along some or all dimensions, but they are not.

SizeElementsMismatchException :: (Index ix, Index ix') => !(Sz ix) -> !(Sz ix') -> SizeException
Total number of elements does not match between the two sizes.

SizeSubregionException :: Index ix => !(Sz ix) -> !ix -> !(Sz ix) -> SizeException
Described subregion is too big for the specified size.

SizeEmptyException :: Index ix => !(Sz ix) -> SizeException
An array with the size cannot contain any elements.

SizeOverflowException :: Index ix => !(Sz ix) -> SizeException
Total number of elements is too large resulting in overflow.

Since: 0.6.0

SizeNegativeException :: Index ix => !(Sz ix) -> SizeException
At least one dimensions contain a negative value.

Since: 0.6.0

data ShapeException

Exception that can happen upon conversion of a ragged type array into the rectangular kind. Which means conversion from lists is susceptible to this exception.

Since: 0.3.0

Constructors

DimTooShortException !Dim !(Sz Ix1) !(Sz Ix1)
Across a specific dimension there was not enough elements for the supplied size

DimTooLongException !Dim !(Sz Ix1) !(Sz Ix1)
Across a specific dimension there was too many elements for the supplied size

ShapeNonEmpty
Expected an empty size, but the shape was not empty.

throwImpossible :: HasCallStack => Exception e => e -> a

Throw an impossible error.

Since: 0.5.6

throwEither :: HasCallStack => Either SomeException a -> a

Throw an error on Left or produce the result on Right. Exception type is lost, so do not expect to be able to catch it as such. Stick to IO if you need exception control flow.

Since: 0.5.6

data Uninitialized

An error that gets thrown when an unitialized element of a boxed array gets accessed. Can only happen when array was constructed with unsafeNew.

Constructors

Uninitialized

guardNumberOfElements :: (MonadThrow m, Index ix, Index ix') => Sz ix -> Sz ix' -> m ()

Throw SizeElementsMismatchException whenever number of elements in both sizes do not match.

Since: 0.3.5

class (Typeable e, Show e) => Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException :: e -> SomeException

fromException :: SomeException -> Maybe e

displayException :: e -> String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

Stateful Monads

class MonadIO m => MonadUnliftIO (m :: Type -> Type)

Monads which allow their actions to be run in IO.

While MonadIO allows an IO action to be lifted into another monad, this class captures the opposite concept: allowing you to capture the monadic context. Note that, in order to meet the laws given below, the intuition is that a monad must have no monadic state, but may have monadic context. This essentially limits MonadUnliftIO to ReaderT and IdentityT transformers on top of IO.

Laws. For any function run provided by withRunInIO, it must meet the monad transformer laws as reformulated for MonadUnliftIO:

run . return = return

run (m >>= f) = run m >>= run . f

Instances of MonadUnliftIO must also satisfy the following laws:

Identity law

withRunInIO (\run -> run m) = m

Inverse law

withRunInIO (\_ -> m) = liftIO m

As an example of an invalid instance, a naive implementation of MonadUnliftIO (StateT s m) might be

withRunInIO inner =
  StateT $ \s ->
    withRunInIO $ \run ->
      inner (run . flip evalStateT s)

This breaks the identity law because the inner run m would throw away any state changes in m.

Since: unliftio-core-0.1.0.0

Minimal complete definition

withRunInIO

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3

class Monad m => PrimMonad (m :: Type -> Type)

Class of monads which can perform primitive state-transformer actions.

Minimal complete definition

primitive

Associated Types

type PrimState (m :: Type -> Type)

State token type.