Data.Massiv.Core.List

Description

Documentation

data L

Constructors

L

data family Array r ix e :: Type

The array family. Representations r describe how data is arranged or computed. All arrays have a common property that each index ix always maps to the same unique element e, even if that element does not yet exist in memory and the array has to be computed in order to get the value of that element. Data is always arranged in a nested row-major fashion. Rank of an array is specified by Dimensions ix.

Since: 0.1.0

newtype List ix e

Constructors

List

Fields

unList :: [Elt ix e]

toListArray :: (Ragged L ix e, Shape r ix, Source r e) => Array r ix e -> Array L ix e

Construct an array backed by linked lists from any source array

Since: 0.4.0

showsArrayPrec

Arguments

:: forall r r' ix e. (Ragged L ix e, Load r ix e, Load r' ix e, Source r' e, Show e)

=> (Array r ix e -> Array r' ix e)
Modifier

-> Int

-> Array r ix e

-> ShowS

Helper function for declaring Show instances for arrays

Since: 0.4.0

showArrayList :: Show arr => [arr] -> String -> String

Helper function for declaring Show instances for arrays

Since: 0.4.0

type family ListItem ix e :: Type where ...

Equations

ListItem Ix1 e = e

ListItem ix e = [ListItem (Lower ix) e]