Data.Massiv.Core.Operations

Description

Documentation

class (Size r, Num e) => FoldNumeric r e where

Minimal complete definition

foldArray, powerSumArray, unsafeDotProduct

Methods

sumArray :: Index ix => Array r ix e -> e

Compute sum of all elements in the array

Since: 0.5.6

productArray :: Index ix => Array r ix e -> e

Compute product of all elements in the array

Since: 0.5.6

powerSumArray :: Index ix => Array r ix e -> Int -> e

Raise each element in the array to some non-negative power and sum the results

Since: 0.5.7

unsafeDotProduct :: Index ix => Array r ix e -> Array r ix e -> e

Compute dot product without any extraneous checks

Since: 0.5.6

foldArray :: Index ix => (e -> e -> e) -> e -> Array r ix e -> e

Fold over an array

Since: 0.5.6

defaultPowerSumArray :: (Index ix, Source r e, Num e) => Array r ix e -> Int -> e

defaultUnsafeDotProduct :: (Num e, Index ix, Source r e) => Array r ix e -> Array r ix e -> e

defaultFoldArray :: (Index ix, Source r e) => (e -> e -> e) -> e -> Array r ix e -> e

class FoldNumeric r e => Numeric r e where

Minimal complete definition

unsafeLiftArray, unsafeLiftArray2

Methods

plusScalar :: Index ix => Array r ix e -> e -> Array r ix e

minusScalar :: Index ix => Array r ix e -> e -> Array r ix e

scalarMinus :: Index ix => e -> Array r ix e -> Array r ix e

multiplyScalar :: Index ix => Array r ix e -> e -> Array r ix e

absPointwise :: Index ix => Array r ix e -> Array r ix e

additionPointwise :: Index ix => Array r ix e -> Array r ix e -> Array r ix e

subtractionPointwise :: Index ix => Array r ix e -> Array r ix e -> Array r ix e

multiplicationPointwise :: Index ix => Array r ix e -> Array r ix e -> Array r ix e

powerPointwise :: Index ix => Array r ix e -> Int -> Array r ix e

Raise each element of the array to the power

unsafeLiftArray :: Index ix => (e -> e) -> Array r ix e -> Array r ix e

unsafeLiftArray2 :: Index ix => (e -> e -> e) -> Array r ix e -> Array r ix e -> Array r ix e

defaultUnsafeLiftArray :: (Load r ix e, Source r e) => (e -> e) -> Array r ix e -> Array r ix e

defaultUnsafeLiftArray2 :: (Load r ix e, Source r e) => (e -> e -> e) -> Array r ix e -> Array r ix e -> Array r ix e

class (Numeric r e, Floating e) => NumericFloat r e where

Minimal complete definition

Nothing

Methods

divideScalar :: Index ix => Array r ix e -> e -> Array r ix e

scalarDivide :: Index ix => e -> Array r ix e -> Array r ix e

divisionPointwise :: Index ix => Array r ix e -> Array r ix e -> Array r ix e

recipPointwise :: Index ix => Array r ix e -> Array r ix e

sqrtPointwise :: Index ix => Array r ix e -> Array r ix e