Data.Massiv.Array.Manifest

Description

Manifest

class Source r ix e => Manifest r ix e

Manifest arrays are backed by actual memory and values are looked up versus computed as it is with delayed arrays. Because of this fact indexing functions (!), (!?), etc. are constrained to manifest arrays only.

Minimal complete definition

unsafeLinearIndexM

toManifest :: Manifest r ix e => Array r ix e -> Array M ix e

O(1) - Conversion of Manifest arrays to M representation.

data M

General Manifest representation

Boxed

data B

Array representation for Boxed elements. This structure is element and spine strict, but elements are strict to Weak Head Normal Form (WHNF) only.

Constructors

B

data BL

Array representation for Boxed elements. This data structure is lazy with respect to its elements, but is strict with respect to the spine.

Constructors

BL

data BN

Array representation for Boxed elements. This structure is element and spine strict, and elements are always in Normal Form (NF), therefore NFData instance is required.

Constructors

BN

type N = BN

Type and pattern N have been added for backwards compatibility and will be replaced in the future in favor of BN

pattern N :: N

data Uninitialized

An error that gets thrown when an unitialized element of a boxed array gets accessed. Can only happen when array was constructed with unsafeNew.

Constructors

Uninitialized

Access

findIndex :: Manifest r ix e => (e -> Bool) -> Array r ix e -> Maybe ix

O(n) - Perform a row-major search starting at 0 for an element. Returns the index of the first occurance of an element or Nothing if a predicate could not be satisifed after it was applyied to all elements of the array.

Since: 0.5.5

Conversion

Important part of all conversions in this section is that the actual boxed Array, which holds the pointers to values isn't copied around, it is always kept as the same array. Conversion to Massiv boxed array will undergo evaluation during which computation strategies will be respected.

toLazyArray :: Array B ix e -> Array BL ix e

O(1) - Cast a strict boxed array into a lazy boxed array.

Since: 0.6.0

evalLazyArray :: Index ix => Array BL ix e -> Array B ix e

O(n) - Evaluate all elements of a boxed lazy array to weak head normal form

Since: 0.6.0

forceLazyArray :: (NFData e, Index ix) => Array BL ix e -> Array N ix e

O(n) - Evaluate all elements of a boxed lazy array to normal form

Since: 0.6.0

unwrapNormalForm :: Array N ix e -> Array B ix e

O(1) - Converts array from N to B representation.

Since: 0.5.0

evalNormalForm :: (Index ix, NFData e) => Array B ix e -> Array N ix e

O(n) - Compute all elements of a boxed array to NF (normal form)

Since: 0.5.0

Primitive Boxed Array

unwrapLazyArray :: Array BL ix e -> Array e

O(1) - Unwrap boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.6.0

wrapLazyArray :: Array e -> Vector BL e

O(1) - Wrap a boxed array.

Since: 0.6.0

unwrapArray :: Array B ix e -> Array e

O(1) - Unwrap boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

evalArray

Arguments

:: Comp
Computation strategy

-> Array e
Lazy boxed array from primitive package.

-> Vector B e

O(n) - Wrap a boxed array and evaluate all elements to a WHNF.

Since: 0.2.1

unwrapMutableArray :: MArray s B ix e -> MutableArray s e

O(1) - Unwrap mutable boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

unwrapMutableLazyArray :: MArray s BL ix e -> MutableArray s e

O(1) - Unwrap mutable boxed lazy array. This will discard any possible slicing that has been applied to the array.

Since: 0.6.0

evalMutableArray

Arguments

:: PrimMonad m

=> MutableArray (PrimState m) e
Mutable array that will get wrapped

-> m (MArray (PrimState m) B Ix1 e)

O(n) - Wrap mutable boxed array and evaluate all elements to WHNF.

Since: 0.2.1

unwrapNormalFormArray :: Array N ix e -> Array e

O(1) - Unwrap a fully evaluated boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

evalNormalFormArray

Arguments

:: NFData e

=> Comp
Computation strategy

-> Array e
Lazy boxed array

-> Array N Ix1 e

O(n) - Wrap a boxed array and evaluate all elements to a Normal Form (NF).

Since: 0.2.1

unwrapNormalFormMutableArray :: MArray s N ix e -> MutableArray s e

O(1) - Unwrap a fully evaluated mutable boxed array. This will discard any possible slicing that has been applied to the array.

Since: 0.2.1

evalNormalFormMutableArray :: (PrimMonad m, NFData e) => MutableArray (PrimState m) e -> m (MArray (PrimState m) N Ix1 e)

O(n) - Wrap mutable boxed array and evaluate all elements to NF.

Since: 0.2.1

Boxed Vector

toBoxedVector :: Index ix => Array BL ix a -> Vector a

O(1) - Converts a boxed Array into a Vector.

Since: 0.5.0

toBoxedMVector :: Index ix => MArray s BL ix a -> MVector s a

O(1) - Converts a boxed MArray into a MVector.

Since: 0.5.0

fromBoxedVector :: Vector a -> Vector BL a

O(1) - Cast a boxed vector without touching any elements.

Since: 0.6.0

fromBoxedMVector :: MVector s a -> MArray s BL Ix1 a

O(1) - Convert mutable boxed vector to a lazy mutable boxed array. Both keep pointing to the same memory

Since: 0.6.0

evalBoxedVector :: Comp -> Vector a -> Array B Ix1 a

O(n) - Convert a boxed vector and evaluate all elements to WHNF. Computation strategy will be respected during evaluation

Since: 0.5.0

evalBoxedMVector :: PrimMonad m => MVector (PrimState m) a -> m (MArray (PrimState m) B Ix1 a)

O(n) - Convert mutable boxed vector and evaluate all elements to WHNF sequentially. Both keep pointing to the same memory

Since: 0.5.0

Primitive

data P

Representation for Primitive elements

Constructors

P

class Prim a

Class of types supporting primitive array operations. This includes interfacing with GC-managed memory (functions suffixed with ByteArray#) and interfacing with unmanaged memory (functions suffixed with Addr#). Endianness is platform-dependent.

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

Conversion

Primitive ByteArray

toByteArray :: (Index ix, Prim e) => Array P ix e -> ByteArray

O(n) - Ensure that the size matches the internal ByteArray. If not make a copy of the slice and return it as ByteArray

Since: 0.2.1

toByteArrayM :: (Prim e, Index ix, MonadThrow m) => Array P ix e -> m ByteArray

O(1) - Unwrap Ensure that the size matches the internal ByteArray.

Since: 0.5.0

unwrapByteArray :: Array P ix e -> ByteArray

O(1) - Extract the internal ByteArray. This will ignore any possible slicing that has been applied to the array. Use toByteArray in order to preserve slicing or unwrapByteArrayOffset to get ahold of the offset

Since: 0.5.0

unwrapByteArrayOffset :: Array P ix e -> Int

O(1) - Extract potential linear offset into the underlying ByteArray, which can also be extracted with unwrapByteArray.

Since: 0.5.9

fromByteArray :: forall e. Prim e => Comp -> ByteArray -> Array P Ix1 e

O(1) - Construct a flat Array from ByteArray

Since: 0.4.0

fromByteArrayM :: (MonadThrow m, Index ix, Prim e) => Comp -> Sz ix -> ByteArray -> m (Array P ix e)

O(1) - Construct a primitive array from the ByteArray. Will return Nothing if number of elements doesn't match.

Since: 0.3.0

fromByteArrayOffsetM :: (MonadThrow m, Index ix, Prim e) => Comp -> Sz ix -> Int -> ByteArray -> m (Array P ix e)

O(1) - Construct a primitive array from the ByteArray. Will return Nothing if number of elements doesn't match.

Since: 0.5.9

toMutableByteArray :: forall ix e m. (Prim e, Index ix, PrimMonad m) => MArray (PrimState m) P ix e -> m (Bool, MutableByteArray (PrimState m))

O(n) - Try to cast a mutable array to MutableByteArray, if sizes do not match make a copy. Returns True if an array was converted without a copy, in which case it means that the source at the resulting array are still pointing to the same location in memory.

Since: 0.5.0

unwrapMutableByteArray :: MArray s P ix e -> MutableByteArray s

O(1) - Extract the internal MutableByteArray. This will discard any possible slicing that has been applied to the array.

Since: 0.5.0

unwrapMutableByteArrayOffset :: MArray s P ix e -> Int

O(1) - Extract the linear offset into underlying MutableByteArray, which can aslo be extracted with unwrapMutableByteArray.

Since: 0.5.9

fromMutableByteArray :: forall e s. Prim e => MutableByteArray s -> MArray s P Ix1 e

O(1) - Construct a flat Array from MutableByteArray

Since: 0.4.0

fromMutableByteArrayM :: (MonadThrow m, Index ix, Prim e) => Sz ix -> MutableByteArray s -> m (MArray s P ix e)

O(1) - Construct a primitive mutable array from the MutableByteArray. Will throw SizeElementsMismatchException if number of elements doesn't match.

Since: 0.3.0

fromMutableByteArrayOffsetM :: (MonadThrow m, Index ix, Prim e) => Sz ix -> Ix1 -> MutableByteArray s -> m (MArray s P ix e)

O(1) - Construct a primitive mutable array from the MutableByteArray. Will throw SizeElementsMismatchException if number of elements doesn't match.

Since: 0.5.9

Primitive Vector

toPrimitiveVector :: Index ix => Array P ix e -> Vector e

O(1) - Cast a primitive array to a primitive vector.

Since: 0.5.0

toPrimitiveMVector :: Index ix => MArray s P ix e -> MVector s e

O(1) - Cast a mutable primitive array to a mutable primitive vector.

Since: 0.5.0

fromPrimitiveVector :: Vector e -> Array P Ix1 e

O(1) - Cast a primitive vector to a primitive array.

Since: 0.5.0

fromPrimitiveMVector :: MVector s e -> MArray s P Ix1 e

O(1) - Cast a mutable primitive vector to a mutable primitive array.

Since: 0.5.0

Storable

data S

Representation for Storable elements

Constructors

S

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

mallocCompute :: forall r ix e. (Source r ix e, Storable e) => Array r ix e -> IO (Array S ix e)

Very similar to computeAs S except load the source array into memory allocated with malloc on C heap. It can potentially be useful when iteroperating with some C programs.

Since: 0.5.9

mallocCopy :: forall ix e. (Index ix, Storable e) => Array S ix e -> IO (Array S ix e)

Allocate memory on C heap with malloc and copy the source array over.

Since: 0.5.9

Conversion

Storable Vector

toStorableVector :: Array S ix e -> Vector e

O(1) - Unwrap storable array and pull out the underlying storable vector.

Since: 0.2.1

toStorableMVector :: MArray s S ix e -> MVector s e

O(1) - Unwrap storable mutable array and pull out the underlying storable mutable vector.

Since: 0.2.1

fromStorableVector :: Storable e => Comp -> Vector e -> Array S Ix1 e

O(1) - Cast a storable vector to a storable array.

Since: 0.5.0

fromStorableMVector :: MVector s e -> MArray s S Ix1 e

O(1) - Cast a mutable storable vector to a mutable storable array.

Since: 0.5.0

Direct Pointer Access

withPtr :: (MonadUnliftIO m, Storable a) => MArray RealWorld S ix a -> (Ptr a -> m b) -> m b

A pointer to the beginning of the mutable array.

Since: 0.1.3

Unboxed

data U

Representation for Unboxed elements

Constructors

U

class (Vector Vector a, MVector MVector a) => Unbox a

Conversion

Unboxed Vector

toUnboxedVector :: Array U ix e -> Vector e

O(1) - Unwrap unboxed array and pull out the underlying unboxed vector.

Since: 0.2.1

toUnboxedMVector :: MArray s U ix e -> MVector s e

O(1) - Unwrap unboxed mutable array and pull out the underlying unboxed mutable vector.

Since: 0.2.1

fromUnboxedVector :: Unbox e => Comp -> Vector e -> Array U Ix1 e

O(1) - Wrap an unboxed vector and produce an unboxed flat array.

Since: 0.6.0

fromUnboxedMVector :: Unbox e => MVector s e -> MArray s U Ix1 e

O(1) - Wrap an unboxed mutable vector and produce a mutable unboxed flat array.

Since: 0.5.0

ByteString Conversion

fromByteString

Arguments

:: Comp
Computation strategy

-> ByteString
Strict ByteString to use as a source.

-> Array M Ix1 Word8

O(1) - Convert a strict ByteString into a manifest array. Will return Nothing if length doesn't match the total number of elements of new array.

Since: 0.2.1

castFromByteString :: Comp -> ByteString -> Array S Ix1 Word8

O(1) - Cast a strict ByteString into a Storable array

Since: 0.3.0

toByteString

Arguments

:: Load r ix Word8

=> Array r ix Word8
Source array

-> ByteString

O(n) - Convert any source array into a strict ByteString. In case when the source array is actually storable, no memory copy will occur.

Since: 0.2.1

castToByteString :: Array S ix Word8 -> ByteString

O(1) - Cast a Storable array into a strict ByteString

Since: 0.3.0

toBuilder :: Source r ix e => (e -> Builder) -> Array r ix e -> Builder

O(n) - Conversion of array monoidally into a ByteString Builder.

Since: 0.2.1

castToBuilder :: Array S ix Word8 -> Builder

O(1) - Cast a storable array of Word8 to ByteString Builder.

Since: 0.5.0