Data.Massiv.Core

Description

Documentation

data family Array r ix e :: *

The array family. Representations r describe how data is arranged or computed. All arrays have a common property that each index ix always maps to the same unique element, even if that element does not yet exist in memory and the arry has to be computed in order to get access to that element. Data is always arranged in a nested row-major fashion, depth of which is controlled by Rank ix.

type Vector r e = Array r Ix1 e

Type synonym for a single dimension array, or simply a flat vector.

Since: 0.5.0

type MVector s r e = MArray s r Ix1 e

Type synonym for a single dimension mutable array, or simply a flat mutable vector.

Since: 0.5.0

type Matrix r e = Array r Ix2 e

Type synonym for a two-dimentsional array, or simply a matrix.

Since: 0.5.0

type MMatrix s r e = MArray s r Ix2 e

Type synonym for a two-dimentsional mutable array, or simply a mutable matrix.

Since: 0.5.0

type family Elt r ix e :: * where ...

Equations

Elt r Ix1 e = e

Elt r ix e = Array (R r) (Lower ix) e

class (Typeable r, Index ix) => Construct r ix e

Array types that can be constructed.

Minimal complete definition

setComp, (makeArray | makeArrayLinear)

class (Typeable r, Index ix) => Load r ix e where

Any array that can be computed and loaded into memory

Minimal complete definition

getComp, size, (loadArrayM | loadArrayWithSetM)

Associated Types

type R r :: *

type R r = r

Methods

loadArrayM

Arguments

:: Monad m

=> Scheduler m ()

-> Array r ix e
Array that is being loaded

-> (Int -> e -> m ())
Function that writes an element into target array

-> m ()

Load an array into memory.

Since: 0.3.0

loadArrayWithSetM

Arguments

:: Monad m

=> Scheduler m ()

-> Array r ix e
Array that is being loaded

-> (Ix1 -> e -> m ())
Function that writes an element into target array

-> (Ix1 -> Sz1 -> e -> m ())
Function that efficiently sets a region of an array to the supplied value target array

-> m ()

Load an array into memory, just like loadArrayM. Except it also accepts a function that is potentially optimized for setting many cells in a region to the same value

Since: 0.5.8

class Load r ix e => Stream r ix e where

Methods

toStream :: Array r ix e -> Steps Id e

toStreamIx :: Array r ix e -> Steps Id (ix, e)

class (Resize r ix, Load r ix e) => Source r ix e

Arrays that can be used as source to practically any manipulation function.

Minimal complete definition

(unsafeIndex | unsafeLinearIndex), unsafeLinearSlice

class Index ix => Resize r ix

Minimal complete definition

unsafeResize

class Load r ix e => Extract r ix e

Minimal complete definition

unsafeExtract

class Load r ix e => StrideLoad r ix e where

Minimal complete definition

Nothing

Methods

loadArrayWithStrideM

Arguments

:: Monad m

=> Scheduler m ()

-> Stride ix
Stride to use

-> Sz ix
Size of the target array affected by the stride.

-> Array r ix e
Array that is being loaded

-> (Int -> e -> m ())
Function that writes an element into target array

-> m ()

Load an array into memory with stride. Default implementation requires an instance of Source.

default loadArrayWithStrideM :: (Source r ix e, Monad m) => Scheduler m () -> Stride ix -> Sz ix -> Array r ix e -> (Int -> e -> m ()) -> m ()

class Load r ix e => Slice r ix e

Minimal complete definition

unsafeSlice

class Load r ix e => OuterSlice r ix e

Minimal complete definition

unsafeOuterSlice

class Load r ix e => InnerSlice r ix e

Minimal complete definition

unsafeInnerSlice

class Source r ix e => Manifest r ix e

Manifest arrays are backed by actual memory and values are looked up versus computed as it is with delayed arrays. Because of this fact indexing functions (!), (!?), etc. are constrained to manifest arrays only.

Minimal complete definition

unsafeLinearIndexM

class (Construct r ix e, Manifest r ix e) => Mutable r ix e

Minimal complete definition

msize, unsafeThaw, unsafeFreeze, unsafeNew, unsafeLinearRead, unsafeLinearWrite, initialize

class Construct r ix e => Ragged r ix e

Minimal complete definition

emptyR, isNull, consR, unconsR, generateRaggedM, edgeSize, flattenRagged, loadRagged, raggedFormat

class Nested r ix e where

Methods

fromNested :: NestedStruct r ix e -> Array r ix e

toNested :: Array r ix e -> NestedStruct r ix e

type family NestedStruct r ix e :: *

data L

Constructors

L

data LN

type family ListItem ix e :: * where ...

Equations

ListItem Ix1 e = e

ListItem ix e = [ListItem (Lower ix) e]

data Scheduler (m :: Type -> Type) a

Main type for scheduling work. See withScheduler or withScheduler_ for ways to construct and use this data type.

Since: scheduler-1.0.0

data SchedulerWS s (m :: Type -> Type) a

This is a wrapper around Scheduler, but it also keeps a separate state for each individual worker. See withSchedulerWS or withSchedulerWS_ for ways to construct and use this data type.

Since: scheduler-1.4.0

data Comp

Computation strategy to use when scheduling work.

Constructors

Seq
Sequential computation

ParOn ![Int]
Schedule workers to run on specific capabilities. Specifying an empty list ParOn [] or using Par will result in utilization of all available capabilities.

ParN !Word16
Specify the number of workers that will be handling all the jobs. Difference from ParOn is that workers can jump between cores. Using ParN 0 will result in using all available capabilities.

Bundled Patterns

pattern Par :: Comp
Parallel computation using all available cores. Same as ParOn []

Since: scheduler-1.0.0

pattern Par' :: Comp
Parallel computation using all available cores. Same as ParN 0

Since: scheduler-1.1.0

appComp :: (Construct r ix e, Load r ix e) => Comp -> Array r ix e -> Array r ix e

Append computation strategy using Comp's Monoid instance.

Since: 0.6.0

data WorkerStates s

Each worker is capable of keeping it's own state, that can be share for different schedulers, but not at the same time. In other words using the same WorkerStates on withSchedulerS concurrently will result in an error. Can be initialized with initWorkerStates

Since: scheduler-1.4.0

initWorkerStates :: MonadIO m => Comp -> (WorkerId -> m s) -> m (WorkerStates s)

Initialize a separate state for each worker.

Since: scheduler-1.4.0

module Data.Massiv.Core.Index

Numeric

class Num e => FoldNumeric r e

Minimal complete definition

powerSumArray, unsafeDotProduct, foldArray

class FoldNumeric r e => Numeric r e

Minimal complete definition

unsafeLiftArray, unsafeLiftArray2

class (Numeric r e, Floating e) => NumericFloat r e

Exceptions

class Monad m => MonadThrow (m :: Type -> Type) where

A class for monads in which exceptions may be thrown.

Instances should obey the following law:

throwM e >> x = throwM e

In other words, throwing an exception short-circuits the rest of the monadic computation.

Methods

throwM :: Exception e => e -> m a

Throw an exception. Note that this throws when this action is run in the monad m, not when it is applied. It is a generalization of Control.Exception's throwIO.

Should satisfy the law:

throwM e >> f = throwM e

throw :: forall (r :: RuntimeRep) (a :: TYPE r) e. Exception e => e -> a

Throw an exception. Exceptions may be thrown from purely functional code, but may only be caught within the IO monad.

data IndexException where

Exceptions that get thrown when there is a problem with an index, size or dimension.

Since: 0.3.0

Constructors

IndexZeroException :: Index ix => !ix -> IndexException
Index contains a zero value along one of the dimensions.

IndexDimensionException :: (NFData ix, Show ix, Typeable ix) => !ix -> !Dim -> IndexException
Dimension is out of reach.

IndexOutOfBoundsException :: Index ix => !(Sz ix) -> !ix -> IndexException
Index is out of bounds.

data SizeException where

Exception that indicates an issue with an array size.

Since: 0.3.0

Constructors

SizeMismatchException :: Index ix => !(Sz ix) -> !(Sz ix) -> SizeException
Two sizes are expected to be equal along some or all dimensions, but they are not.

SizeElementsMismatchException :: (Index ix, Index ix') => !(Sz ix) -> !(Sz ix') -> SizeException
Total number of elements does not match between the two sizes.

SizeSubregionException :: Index ix => !(Sz ix) -> !ix -> !(Sz ix) -> SizeException
Described subregion is too big for the specified size.

SizeEmptyException :: Index ix => !(Sz ix) -> SizeException
An array with the size cannot contain any elements.

SizeOverflowException :: Index ix => !(Sz ix) -> SizeException
Total number of elements is too large resulting in overflow.

Since: 0.6.0

SizeNegativeException :: Index ix => !(Sz ix) -> SizeException
At least one dimensions contain a negative value.

Since: 0.6.0

data ShapeException

Exception that can happen upon conversion of a ragged type array into the rectangular kind. Which means conversion from lists is susceptible to this exception.

Since: 0.3.0

Constructors

DimTooShortException !Sz1 !Sz1

DimTooLongException

newtype ImpossibleException

Constructors

ImpossibleException SomeException

throwImpossible :: Exception e => e -> a

throwEither :: Either SomeException a -> a

data Uninitialized

An error that gets thrown when an unitialized element of a boxed array gets accessed. Can only happen when array was constructed with unsafeNew.

Constructors

Uninitialized

guardNumberOfElements :: (MonadThrow m, Index ix, Index ix') => Sz ix -> Sz ix' -> m ()

Throw SizeElementsMismatchException whenever number of elements in both sizes do not match.

Since: 0.3.5

class (Typeable e, Show e) => Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException :: e -> SomeException

fromException :: SomeException -> Maybe e

displayException :: e -> String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

Stateful Monads

class MonadIO m => MonadUnliftIO (m :: Type -> Type)

Monads which allow their actions to be run in IO.

While MonadIO allows an IO action to be lifted into another monad, this class captures the opposite concept: allowing you to capture the monadic context. Note that, in order to meet the laws given below, the intuition is that a monad must have no monadic state, but may have monadic context. This essentially limits MonadUnliftIO to ReaderT and IdentityT transformers on top of IO.

Laws. For any value u returned by askUnliftIO, it must meet the monad transformer laws as reformulated for MonadUnliftIO:

unliftIO u . return = return

unliftIO u (m >>= f) = unliftIO u m >>= unliftIO u . f

Instances of MonadUnliftIO must also satisfy the idempotency law:

askUnliftIO >>= \u -> (liftIO . unliftIO u) m = m

This law showcases two properties. First, askUnliftIO doesn't change the monadic context, and second, liftIO . unliftIO u is equivalent to id IF called in the same monadic context as askUnliftIO.

Since: unliftio-core-0.1.0.0

Minimal complete definition

withRunInIO

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

class Monad m => PrimMonad (m :: Type -> Type)

Class of monads which can perform primitive state-transformer actions

Minimal complete definition

primitive

Associated Types

type PrimState (m :: Type -> Type)

State token type