Data.Massiv.Core.List

Description

Documentation

data LN

data L

Constructors

L

data family Array r ix e :: *

The array family. Representations r describe how data is arranged or computed. All arrays have a common property that each index ix always maps to the same unique element, even if that element does not yet exist in memory and the arry has to be computed in order to get access to that element. Data is always arranged in a nested row-major fashion, depth of which is controlled by Rank ix.

toListArray :: (Construct L ix e, Source r ix e) => Array r ix e -> Array L ix e

Construct an array backed by linked lists from any source array

Since: 0.4.0

showsArrayPrec

Arguments

:: forall r r' ix ix' e. (Ragged L ix' e, Load r ix e, Source r' ix' e, Show e)

=> (Array r ix e -> Array r' ix' e)
Modifier

-> Int

-> Array r ix e

-> ShowS

Helper function for declaring Show instances for arrays

Since: 0.4.0

showArrayList :: Show arr => [arr] -> String -> String

Helper function for declaring Show instances for arrays

Since: 0.4.0

type family ListItem ix e :: * where ...

Equations

ListItem Ix1 e = e

ListItem ix e = [ListItem (Lower ix) e]